Fix off-by-one
Well, something was off by one. The database function
populate_grid creates grids with coordinates 0..n inclusive,
so we need to accommodate that here.

I’m not entirely sure which is correct, to be honest, but
since I tend to think of the grid points as being at the
interstices between the pixels I’m going for this.