Cached region paths
Add new options --dump-regions and --load-regions, which may be
used to add a new map to an existing set or to change an existing
one slightly without recomputing the optimal path simplification.

Also remove the old --include-segments option, since it is not used
at present and it would have needed reimplementing to accommodate
these changes.