Cache interpolators
Even using scipy, loading the cartogram grids and creating interpolators
for them is one of the slowest parts of the cartogram generation process.

Therefore we cache the interpolators using cPickle, and use the cache
if it is newer than the original cart file. This leads to a large
speed improvement.