Fix amazing bug; refactor
Oh, Sweet Mary and all the Saints, what a remarkable bug this was.
A stray trailing comma when breakpoints was assigned made it into
a one-element tuple containing the set of breakpoints, which then
meant that the breakpoints were not at all being detected.

This change also refactors the code to avoid a needless second database
query.

Note that this correction exposes a bug in the simplification code,
that was previously masked by the fact that no breakpoints were being
detected.