Major rework of density grid generation
Rework the generation of density grids for better accuracy. In
particular, use a two-pass process where the values for zero
and missing areas, and for oceans, are filled in on the second
pass.

This allows us to make the ocean density equal to the global average
very closely, so avoiding the effect where the whole world becomes
larger or smaller on certain maps.