Make it work with current PostGIS
The previous logic does not work correctly with the current version
of PostGIS. (I assume it did work previously, since it has been used
without incident.)

The problem is that the envelope of the Robinson map, when converted
to Lat/Lon coordinates, was *not* the whole world. This is because the
corners of the envelope lie outside the Robinson map, and are not mapped
to the corners of the 4326 projection.

The solution implemented here is to do the transformation in the other
direction, from Robinson (or whatever) to 4326, and make the comparison
there. This is less efficient, but also less broken.